Use long

+++ MathSample/PiWpf/PiCalculator.cs
@@ -16,7 +16,7 @@
 			IsOn = true;
 			Pi_d = 0;
 			var pos = true;
-			for (int i = 1; IsOn; i += 2, pos ^= true)
+			for (long i = 1; IsOn; i += 2, pos ^= true)
 			{
 				var d = 4m / i;
 				if (pos) Pi_d += d;
@@ -30,7 +30,7 @@
 			Pi = 0;
 			RealNumber n4 = 4;
 			var pos = true;
-			for (int i = 1; IsOn; i += 2, pos ^= true)
+			for (long i = 1; IsOn; i += 2, pos ^= true)
 			{
 				var d = n4 / i;
 				if (pos) Pi += d;
@@ -44,7 +44,7 @@
 			var r12 = (decimal)Math.Sqrt(12);
 			Pi_d = r12;
 			var p = 1m;
-			for (int i = 3; IsOn; i += 2)
+			for (long i = 3; IsOn; i += 2)
 			{
 				p *= -3;
 				Pi_d += r12 / (i * p);
@@ -57,7 +57,7 @@
 			var r12 = RealNumber.Sqrt(12);
 			Pi = r12;
 			RealNumber p = 1;
-			for (int i = 3; IsOn; i += 2)
+			for (long i = 3; IsOn; i += 2)
 			{
 				p *= -3;
 				Pi += r12 / (i * p);
